requirements.txt
Flask==3.1.3
flask-cors==5.0.0
flask-socketio==5.5.0
eventlet==0.36.1
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
grpcio>=1.54.0
numpy>=1.24.0
lxml>=4.9.0
cryptography>=41.0.0
gunicorn==23.0.0
catboost>=1.2.0
lightgbm>=4.3.0
ta>=0.11.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
https://github.com/rongardF/tvdatafeed/archive/refs/heads/main.zip
